requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
transformers
pinecone-client
pandas
scikit-learn
